requirements.txt
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-addanother==2.0.2
django-admin-thumbnails==0.2.6
django-appconf==1.0.5
django-autocomplete-light==3.8.2
django-autoslug==1.9.8
django-bootstrap4==21.1
django-ckeditor==6.1.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-mptt==0.13.4
django-nested-admin==3.3.3
django-select2==7.9.0
django-six==1.0.4
django-utils-six==2.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==2.10
Markdown==3.3.4
Pillow==8.2.0
pycodestyle==2.7.0
python-decouple==3.4
python-monkey-business==1.0.0
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.3
sqlparse==0.4.1
stripe==2.58.0
toml==0.10.2
Unipath==1.1
urllib3==1.26.5
whitenoise==5.2.0
